fixed bug for duplicated Less (bfloat16) operation registration

+++ tensorflow/core/kernels/cwise_op_less.cc
@@ -19,7 +19,7 @@ namespace tensorflow {
 REGISTER5(BinaryOp, CPU, "Less", functor::less, float, Eigen::half, double,
           bfloat16, int32);
 REGISTER5(BinaryOp, CPU, "Less", functor::less, int64, uint8, int8, int16,
-          bfloat16);
+          uint16);
 
 #if GOOGLE_CUDA
 REGISTER7(BinaryOp, GPU, "Less", functor::less, float, Eigen::half, double,
